reduced size, action filters dash 2
I have successfully reduced the size of the tableau workbook from 60MB to 20MB. I hope this speeds things up. I took out a lot of variables. Based on my preliminary analysis of correlations of variables to food access, I selected only hs graduation, age 65+, percent with public assitance, poverty rate,  and uninsured rate to keep in the dataset. I also scratched all of the scaled variable.. Yuhang, I also nixed your small multiples. It relied on a lot of variables, it didn't make sense to me and I feel like it was redundant with my filter groups. Let me know if you want it back in. I have committed the new dataset and rcode to github. I have also written a new introduction as well as instructions for the webpage which I have placed in a word document on github called "Content for dashboard"